models/compositionNetV1/requirements.txt
scipy==1.5.0
opencv_python_headless==4.4.0.46
tensorflow_gpu==2.4.0
pandas==1.0.5
~umpy==1.18.5
h5py==2.10.0
Keras==2.4.3
numpy==1.21.2
Pillow==8.3.2
PyYAML==5.4.1
tensorflow==2.6.0
